requirements-cpu.txt
certifi==2022.12.7
charset-normalizer==3.1.0
filelock==3.11.0
idna==3.4
Jinja2==3.1.2
llvmlite==0.39.1
MarkupSafe==2.1.2
mpmath==1.3.0
networkx==3.1
numba==0.56.4
numpy==1.23.5
opencv-python==4.6.0.66
Pillow==9.5.0
pygame==2.3.0
requests==2.28.2
sympy==1.11.1
torch==2.0.0
torchvision==0.15.1+cpu
typing_extensions==4.5.0
urllib3==1.26.15
